backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-ninja==0.22.2
install==1.3.5
psycopg2-binary==2.9.7
pydantic==1.10.12
sqlparse==0.4.4
typing-extensions==4.7.1
django-cors-headers
